# Optimization: do not allocate temporary objects in MatrixUtils.transformRect