--- a/packages/flutter/test/painting/matrix_utils_test.dart
+++ b/packages/flutter/test/painting/matrix_utils_test.dart
@@ -121,4 +121,66 @@ void main() {
       forcedOffset,
     );
   });
+
+  test('transformRect with no perspective (w = 1)', () {
+    const Rect l10t20r30b40 = Rect.fromLTRB(10, 20, 30, 40);
+    const Rect l20t40r60b80 = Rect.fromLTRB(20, 40, 60, 80);
+
+    // Identity
+    expect(
+      MatrixUtils.transformRect(Matrix4.identity(), l10t20r30b40),
+      l10t20r30b40,
+    );
+
+    // 2D Scaling
+    expect(
+      MatrixUtils.transformRect(Matrix4.diagonal3Values(2, 2, 2), l10t20r30b40),
+      l20t40r60b80,
+    );
+
+    // Rotation
+    expect(
+      MatrixUtils.transformRect(Matrix4.rotationZ(pi / 2.0), l10t20r30b40),
+      within<Rect>(distance: 0.00001, from: const Rect.fromLTRB(-40.0, 10.0, -20.0, 30.0)),
+    );
+  });
+
+  test('transformRect with perspective (w != 1)', () {
+    final Matrix4 transform = MatrixUtils.createCylindricalProjectionTransform(
+      radius: 10.0,
+      angle: pi / 8.0,
+      perspective: 0.3,
+    );
+
+    for (int i = 1; i < 10000; i++) {
+      final Rect rect = Rect.fromLTRB(11.0 * i, 12.0 * i, 15.0 * i, 18.0 * i);
+      final Rect golden = _vectorWiseTransformRect(transform, rect);
+      expect(MatrixUtils.transformRect(transform, rect), golden);
+    }
+  });
+}
+
+// Produces the same computation as `MatrixUtils.transformRect` but it does this
+// one point at a time. This function is used as the golden implementation of the
+// optimized `MatrixUtils.transformRect` to make sure optimizations do not contain
+// bugs.
+Rect _vectorWiseTransformRect(Matrix4 transform, Rect rect) {
+  final Offset point1 = MatrixUtils.transformPoint(transform, rect.topLeft);
+  final Offset point2 = MatrixUtils.transformPoint(transform, rect.topRight);
+  final Offset point3 = MatrixUtils.transformPoint(transform, rect.bottomLeft);
+  final Offset point4 = MatrixUtils.transformPoint(transform, rect.bottomRight);
+  return Rect.fromLTRB(
+      _min4(point1.dx, point2.dx, point3.dx, point4.dx),
+      _min4(point1.dy, point2.dy, point3.dy, point4.dy),
+      _max4(point1.dx, point2.dx, point3.dx, point4.dx),
+      _max4(point1.dy, point2.dy, point3.dy, point4.dy),
+  );
+}
+
+double _min4(double a, double b, double c, double d) {
+  return min(a, min(b, min(c, d)));
+}
+
+double _max4(double a, double b, double c, double d) {
+  return max(a, max(b, max(c, d)));
 }