## Description

Optimize the performance of `MatrixUtils.transformRect`. This function is called 200-300 times per frame when accessibility is enabled in the material demo page in Flutter Gallery. On Nexus 5X each call costs 2.7 microseconds, which is ~0.8ms of frame time.

The optimization performs the transformation as a single matrix multiplication, and it uses a statically allocated temporary matrix to store intermediate results. This removes 4 `Offset` objects, 4 `Vector3` objects, 4 `Float64List` objects. It replaces 4 separate Matrix x Vector multiplications with one Matrix x Matrix multiplication.

Results:

- Dart AOT (Nexus 5X): 1.42x speed-up (2.7µs => 1.89µs average)
- dart2js (Nexus 5X):  3.09x speed-up (1.52µs => 0.49µs average)
- dart2js (Xeon): 3.84x speed-up (0.62µs => 0.16µs average)

## Related Issues

No issues have been filed for this, but we do observe slowness in Flutter for web when accessibility is enabled. `MatrixUtils.transformRect` shows up in our performance profiles.

## Tests

I added the following tests:

- Added new tests in `matrix_utils_test.dart`

## Checklist

Before you create this PR confirm that it meets all requirements listed below by checking the relevant checkboxes (`[x]`). This will ensure a smooth and quick review process.

- [X] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [X] I signed the [CLA].
- [X] I read and followed the [Flutter Style Guide], including [Features we expect every widget to implement].
- [X] I updated/added relevant documentation (doc comments with `///`).
- [X] All existing and new tests are passing.
- [X] The analyzer (`flutter analyze --flutter-repo`) does not report any problems on my PR.
- [X] I am willing to follow-up on review comments in a timely manner.

## Breaking Change

Does your PR require Flutter developers to manually update their apps to accommodate your change?

- [ ] Yes, this is a breaking change (Please read [Handling breaking changes]). *Replace this with a link to the e-mail where you asked for input on this proposed change.*
- [X] No, this is *not* a breaking change.