@liyuqian 2019-06-27
Let's add a benchmark to the device lab?