@yjbanov 2019-06-27
@liyuqian 

> Let's add a benchmark to the device lab?

Done.