@jim-flar 2019-07-01
I have an alternate fix which is even faster (more than 2x faster than the suggested fix for common cases and just under 2x faster for perspective transforms), but I've only run the benchmark on Android devices so far.  We can talk about how to run the benchmarks on Chrome when Yegor has some free time to run me through the process.

The results I see on two Motorola devices are:

// On Moto E2:
// I/flutter (15259): MatrixUtils.transformRectClassic: 3.2 µs per iteration
// I/flutter (15259): MatrixUtils.transformRectYegor: 2.3 µs per iteration
// I/flutter (15259): MatrixUtils.transformRectFlar: 1.0 µs per iteration
// I/flutter (15259): MatrixUtils.pTransformRectClassic: 3.2 µs per iteration
// I/flutter (15259): MatrixUtils.pTransformRectYegor: 2.3 µs per iteration
// I/flutter (15259): MatrixUtils.pTransformRectFlar: 1.4 µs per iteration

// On Moto G4:
// I/flutter (20474): MatrixUtils.transformRectClassic: 1.9 µs per iteration
// I/flutter (20474): MatrixUtils.transformRectYegor: 1.2 µs per iteration
// I/flutter (20474): MatrixUtils.transformRectFlar: 0.5 µs per iteration
// I/flutter (20474): MatrixUtils.pTransformRectClassic: 1.8 µs per iteration
// I/flutter (20474): MatrixUtils.pTransformRectYegor: 1.3 µs per iteration
// I/flutter (20474): MatrixUtils.pTransformRectFlar: 0.7 µs per iteration