@yjbanov 2019-07-15
Update: I'm back and ready to resume this work. However, @flar may have a more optimal solution, so I may close this PR after catching up with him.