@yjbanov 2019-07-19
Closing in favor of https://github.com/flutter/flutter/pull/36396